Update encoder  * pull `BROTLI_MAX_BACKWARD_LIMIT` to constants  * split generic and Zopfli backward references code  * pull hashers init and stitch invocation to encoder  * make `dictionary_hash` a compilation unit  * add `size hint` parameter  * add new hasher  * use `size hint` to pick new hasher for q4  * modernize clz guard (fix #495)  * move `hash to binary tree` to separate file  * add `Initialize` and `Cleanup` to all hashers  * do not raise OOM if malloc(0) == NULL (fix #500)
