+++ enc/compress_fragment.h
@@ -40,7 +40,7 @@ extern "C" {
    REQUIRES: All elements in "table[0..table_size-1]" are initialized to zero.
    REQUIRES: "table_size" is an odd (9, 11, 13, 15) power of two
    OUTPUT: maximal copy distance <= |input_size|
-   OUTPUT: maximal copy distance <= MaxBackwardLimit(18) */
+   OUTPUT: maximal copy distance <= BROTLI_MAX_BACKWARD_LIMIT(18) */
 BROTLI_INTERNAL void BrotliCompressFragmentFast(MemoryManager* m,
                                                 const uint8_t* input,
                                                 size_t input_size,
