+++ enc/hash.h
@@ -27,8 +27,6 @@
 extern "C" {
 #endif
 
-#define MAX_TREE_SEARCH_DEPTH 64
-#define MAX_TREE_COMP_LENGTH 128
 #define score_t size_t
 
 static const uint32_t kDistanceCacheIndex[] = {
@@ -62,6 +60,7 @@ typedef struct DictionarySearchStatictics {
      for this use.
    * The number has been tuned heuristically against compression benchmarks. */
 static const uint32_t kHashMul32 = 0x1e35a7bd;
+static const uint64_t kHashMul64 = BROTLI_MAKE_UINT64_T(0x1e35a7bd, 0x1e35a7bd);
 
 static BROTLI_INLINE uint32_t Hash14(const uint8_t* data) {
   uint32_t h = BROTLI_UNALIGNED_LOAD32(data) * kHashMul32;
@@ -221,323 +220,17 @@ static BROTLI_INLINE size_t BackwardMatchLengthCode(const BackwardMatch* self) {
 #define CAT(a, b) a ## b
 #define FN(X) EXPAND_CAT(X, HASHER())
 
-#define MAX_NUM_MATCHES_H10 (64 + MAX_TREE_SEARCH_DEPTH)
-
 #define HASHER() H10
-#define HashToBinaryTree HASHER()
-
 #define BUCKET_BITS 17
-#define BUCKET_SIZE (1 << BUCKET_BITS)
-
-static size_t FN(HashTypeLength)(void) { return 4; }
-static size_t FN(StoreLookahead)(void) { return MAX_TREE_COMP_LENGTH; }
-
-static uint32_t FN(HashBytes)(const uint8_t *data) {
-  uint32_t h = BROTLI_UNALIGNED_LOAD32(data) * kHashMul32;
-  /* The higher bits contain more mixture from the multiplication,
-     so we take our results from there. */
-  return h >> (32 - BUCKET_BITS);
-}
-
-/* A (forgetful) hash table where each hash bucket contains a binary tree of
-   sequences whose first 4 bytes share the same hash code.
-   Each sequence is MAX_TREE_COMP_LENGTH long and is identified by its starting
-   position in the input data. The binary tree is sorted by the lexicographic
-   order of the sequences, and it is also a max-heap with respect to the
-   starting positions. */
-typedef struct HashToBinaryTree {
-  /* The window size minus 1 */
-  size_t window_mask_;
-
-  /* Hash table that maps the 4-byte hashes of the sequence to the last
-     position where this hash was found, which is the root of the binary
-     tree of sequences that share this hash bucket. */
-  uint32_t buckets_[BUCKET_SIZE];
-
-  /* The union of the binary trees of each hash bucket. The root of the tree
-     corresponding to a hash is a sequence starting at buckets_[hash] and
-     the left and right children of a sequence starting at pos are
-     forest_[2 * pos] and forest_[2 * pos + 1]. */
-  uint32_t* forest_;
-
-  /* A position used to mark a non-existent sequence, i.e. a tree is empty if
-     its root is at invalid_pos_ and a node is a leaf if both its children
-     are at invalid_pos_. */
-  uint32_t invalid_pos_;
-
-  size_t forest_size_;
-  BROTLI_BOOL is_dirty_;
-} HashToBinaryTree;
-
-static void FN(Reset)(HashToBinaryTree* self) {
-  self->is_dirty_ = BROTLI_TRUE;
-}
-
-static void FN(Initialize)(HashToBinaryTree* self) {
-  self->forest_ = NULL;
-  self->forest_size_ = 0;
-  FN(Reset)(self);
-}
-
-static void FN(Cleanup)(MemoryManager* m, HashToBinaryTree* self) {
-  BROTLI_FREE(m, self->forest_);
-}
-
-static void FN(Init)(
-    MemoryManager* m, HashToBinaryTree* self, const uint8_t* data,
-    const BrotliEncoderParams* params, size_t position, size_t bytes,
-    BROTLI_BOOL is_last) {
-  if (self->is_dirty_) {
-    uint32_t invalid_pos;
-    size_t num_nodes;
-    uint32_t i;
-    BROTLI_UNUSED(data);
-    self->window_mask_ = (1u << params->lgwin) - 1u;
-    invalid_pos = (uint32_t)(0 - self->window_mask_);
-    self->invalid_pos_ = invalid_pos;
-    for (i = 0; i < BUCKET_SIZE; i++) {
-      self->buckets_[i] = invalid_pos;
-    }
-    num_nodes = (position == 0 && is_last) ? bytes : self->window_mask_ + 1;
-    if (num_nodes > self->forest_size_) {
-      BROTLI_FREE(m, self->forest_);
-      self->forest_ = BROTLI_ALLOC(m, uint32_t, 2 * num_nodes);
-      if (BROTLI_IS_OOM(m)) return;
-      self->forest_size_ = num_nodes;
-    }
-    self->is_dirty_ = BROTLI_FALSE;
-  }
-}
-
-static BROTLI_INLINE size_t FN(LeftChildIndex)(HashToBinaryTree* self,
-    const size_t pos) {
-  return 2 * (pos & self->window_mask_);
-}
-
-static BROTLI_INLINE size_t FN(RightChildIndex)(HashToBinaryTree* self,
-    const size_t pos) {
-  return 2 * (pos & self->window_mask_) + 1;
-}
-
-/* Stores the hash of the next 4 bytes and in a single tree-traversal, the
-   hash bucket's binary tree is searched for matches and is re-rooted at the
-   current position.
-
-   If less than MAX_TREE_COMP_LENGTH data is available, the hash bucket of the
-   current position is searched for matches, but the state of the hash table
-   is not changed, since we can not know the final sorting order of the
-   current (incomplete) sequence.
-
-   This function must be called with increasing cur_ix positions. */
-static BROTLI_INLINE BackwardMatch* FN(StoreAndFindMatches)(
-    HashToBinaryTree* self, const uint8_t* const BROTLI_RESTRICT data,
-    const size_t cur_ix, const size_t ring_buffer_mask, const size_t max_length,
-    const size_t max_backward, size_t* const BROTLI_RESTRICT best_len,
-    BackwardMatch* BROTLI_RESTRICT matches) {
-  const size_t cur_ix_masked = cur_ix & ring_buffer_mask;
-  const size_t max_comp_len =
-      BROTLI_MIN(size_t, max_length, MAX_TREE_COMP_LENGTH);
-  const BROTLI_BOOL should_reroot_tree =
-      TO_BROTLI_BOOL(max_length >= MAX_TREE_COMP_LENGTH);
-  const uint32_t key = FN(HashBytes)(&data[cur_ix_masked]);
-  size_t prev_ix = self->buckets_[key];
-  /* The forest index of the rightmost node of the left subtree of the new
-     root, updated as we traverse and re-root the tree of the hash bucket. */
-  size_t node_left = FN(LeftChildIndex)(self, cur_ix);
-  /* The forest index of the leftmost node of the right subtree of the new
-     root, updated as we traverse and re-root the tree of the hash bucket. */
-  size_t node_right = FN(RightChildIndex)(self, cur_ix);
-  /* The match length of the rightmost node of the left subtree of the new
-     root, updated as we traverse and re-root the tree of the hash bucket. */
-  size_t best_len_left = 0;
-  /* The match length of the leftmost node of the right subtree of the new
-     root, updated as we traverse and re-root the tree of the hash bucket. */
-  size_t best_len_right = 0;
-  size_t depth_remaining;
-  if (should_reroot_tree) {
-    self->buckets_[key] = (uint32_t)cur_ix;
-  }
-  for (depth_remaining = MAX_TREE_SEARCH_DEPTH; ; --depth_remaining) {
-    const size_t backward = cur_ix - prev_ix;
-    const size_t prev_ix_masked = prev_ix & ring_buffer_mask;
-    if (backward == 0 || backward > max_backward || depth_remaining == 0) {
-      if (should_reroot_tree) {
-        self->forest_[node_left] = self->invalid_pos_;
-        self->forest_[node_right] = self->invalid_pos_;
-      }
-      break;
-    }
-    {
-      const size_t cur_len = BROTLI_MIN(size_t, best_len_left, best_len_right);
-      size_t len;
-      assert(cur_len <= MAX_TREE_COMP_LENGTH);
-      len = cur_len +
-          FindMatchLengthWithLimit(&data[cur_ix_masked + cur_len],
-                                   &data[prev_ix_masked + cur_len],
-                                   max_length - cur_len);
-      assert(0 == memcmp(&data[cur_ix_masked], &data[prev_ix_masked], len));
-      if (matches && len > *best_len) {
-        *best_len = len;
-        InitBackwardMatch(matches++, backward, len);
-      }
-      if (len >= max_comp_len) {
-        if (should_reroot_tree) {
-          self->forest_[node_left] =
-              self->forest_[FN(LeftChildIndex)(self, prev_ix)];
-          self->forest_[node_right] =
-              self->forest_[FN(RightChildIndex)(self, prev_ix)];
-        }
-        break;
-      }
-      if (data[cur_ix_masked + len] > data[prev_ix_masked + len]) {
-        best_len_left = len;
-        if (should_reroot_tree) {
-          self->forest_[node_left] = (uint32_t)prev_ix;
-        }
-        node_left = FN(RightChildIndex)(self, prev_ix);
-        prev_ix = self->forest_[node_left];
-      } else {
-        best_len_right = len;
-        if (should_reroot_tree) {
-          self->forest_[node_right] = (uint32_t)prev_ix;
-        }
-        node_right = FN(LeftChildIndex)(self, prev_ix);
-        prev_ix = self->forest_[node_right];
-      }
-    }
-  }
-  return matches;
-}
-
-/* Finds all backward matches of &data[cur_ix & ring_buffer_mask] up to the
-   length of max_length and stores the position cur_ix in the hash table.
-
-   Sets *num_matches to the number of matches found, and stores the found
-   matches in matches[0] to matches[*num_matches - 1]. The matches will be
-   sorted by strictly increasing length and (non-strictly) increasing
-   distance. */
-static BROTLI_INLINE size_t FN(FindAllMatches)(HashToBinaryTree* self,
-    const uint8_t* data, const size_t ring_buffer_mask, const size_t cur_ix,
-    const size_t max_length, const size_t max_backward,
-    const BrotliEncoderParams* params, BackwardMatch* matches) {
-  BackwardMatch* const orig_matches = matches;
-  const size_t cur_ix_masked = cur_ix & ring_buffer_mask;
-  size_t best_len = 1;
-  const size_t short_match_max_backward =
-      params->quality != HQ_ZOPFLIFICATION_QUALITY ? 16 : 64;
-  size_t stop = cur_ix - short_match_max_backward;
-  uint32_t dict_matches[BROTLI_MAX_STATIC_DICTIONARY_MATCH_LEN + 1];
-  size_t i;
-  if (cur_ix < short_match_max_backward) { stop = 0; }
-  for (i = cur_ix - 1; i > stop && best_len <= 2; --i) {
-    size_t prev_ix = i;
-    const size_t backward = cur_ix - prev_ix;
-    if (BROTLI_PREDICT_FALSE(backward > max_backward)) {
-      break;
-    }
-    prev_ix &= ring_buffer_mask;
-    if (data[cur_ix_masked] != data[prev_ix] ||
-        data[cur_ix_masked + 1] != data[prev_ix + 1]) {
-      continue;
-    }
-    {
-      const size_t len =
-          FindMatchLengthWithLimit(&data[prev_ix], &data[cur_ix_masked],
-                                   max_length);
-      if (len > best_len) {
-        best_len = len;
-        InitBackwardMatch(matches++, backward, len);
-      }
-    }
-  }
-  if (best_len < max_length) {
-    matches = FN(StoreAndFindMatches)(self, data, cur_ix, ring_buffer_mask,
-        max_length, max_backward, &best_len, matches);
-  }
-  for (i = 0; i <= BROTLI_MAX_STATIC_DICTIONARY_MATCH_LEN; ++i) {
-    dict_matches[i] = kInvalidMatch;
-  }
-  {
-    size_t minlen = BROTLI_MAX(size_t, 4, best_len + 1);
-    if (BrotliFindAllStaticDictionaryMatches(&data[cur_ix_masked], minlen,
-                                             max_length, &dict_matches[0])) {
-      size_t maxlen = BROTLI_MIN(
-          size_t, BROTLI_MAX_STATIC_DICTIONARY_MATCH_LEN, max_length);
-      size_t l;
-      for (l = minlen; l <= maxlen; ++l) {
-        uint32_t dict_id = dict_matches[l];
-        if (dict_id < kInvalidMatch) {
-          InitDictionaryBackwardMatch(matches++,
-              max_backward + (dict_id >> 5) + 1, l, dict_id & 31);
-        }
-      }
-    }
-  }
-  return (size_t)(matches - orig_matches);
-}
-
-/* Stores the hash of the next 4 bytes and re-roots the binary tree at the
-   current sequence, without returning any matches.
-   REQUIRES: ix + MAX_TREE_COMP_LENGTH <= end-of-current-block */
-static BROTLI_INLINE void FN(Store)(HashToBinaryTree* self, const uint8_t *data,
-    const size_t mask, const size_t ix) {
-  /* Maximum distance is window size - 16, see section 9.1. of the spec. */
-  const size_t max_backward = self->window_mask_ - BROTLI_WINDOW_GAP + 1;
-  FN(StoreAndFindMatches)(self, data, ix, mask, MAX_TREE_COMP_LENGTH,
-      max_backward, NULL, NULL);
-}
-
-static BROTLI_INLINE void FN(StoreRange)(HashToBinaryTree* self,
-    const uint8_t *data, const size_t mask, const size_t ix_start,
-    const size_t ix_end) {
-  size_t i = ix_start;
-  size_t j = ix_start;
-  if (ix_start + 63 <= ix_end) {
-    i = ix_end - 63;
-  }
-  if (ix_start + 512 <= i) {
-    for (; j < i; j += 8) {
-      FN(Store)(self, data, mask, j);
-    }
-  }
-  for (; i < ix_end; ++i) {
-    FN(Store)(self, data, mask, i);
-  }
-}
-
-static BROTLI_INLINE void FN(StitchToPreviousBlock)(HashToBinaryTree* self,
-    size_t num_bytes, size_t position, const uint8_t* ringbuffer,
-    size_t ringbuffer_mask) {
-  if (num_bytes >= FN(HashTypeLength)() - 1 &&
-      position >= MAX_TREE_COMP_LENGTH) {
-    /* Store the last `MAX_TREE_COMP_LENGTH - 1` positions in the hasher.
-       These could not be calculated before, since they require knowledge
-       of both the previous and the current block. */
-    const size_t i_start = position - MAX_TREE_COMP_LENGTH + 1;
-    const size_t i_end = BROTLI_MIN(size_t, position, i_start + num_bytes);
-    size_t i;
-    for (i = i_start; i < i_end; ++i) {
-      /* Maximum distance is window size - 16, see section 9.1. of the spec.
-         Furthermore, we have to make sure that we don't look further back
-         from the start of the next block than the window size, otherwise we
-         could access already overwritten areas of the ring-buffer. */
-      const size_t max_backward =
-          self->window_mask_ - BROTLI_MAX(size_t,
-                                          BROTLI_WINDOW_GAP - 1,
-                                          position - i);
-      /* We know that i + MAX_TREE_COMP_LENGTH <= position + num_bytes, i.e. the
-         end of the current block and that we have at least
-         MAX_TREE_COMP_LENGTH tail in the ring-buffer. */
-      FN(StoreAndFindMatches)(self, ringbuffer, i, ringbuffer_mask,
-          MAX_TREE_COMP_LENGTH, max_backward, NULL, NULL);
-    }
-  }
-}
-
-#undef BUCKET_SIZE
+#define MAX_TREE_SEARCH_DEPTH 64
+#define MAX_TREE_COMP_LENGTH 128
+#include "./hash_to_binary_tree_inc.h"  /* NOLINT(build/include) */
+#undef MAX_TREE_SEARCH_DEPTH
+#undef MAX_TREE_COMP_LENGTH
 #undef BUCKET_BITS
-
 #undef HASHER
+/* MAX_NUM_MATCHES == 64 + MAX_TREE_SEARCH_DEPTH */
+#define MAX_NUM_MATCHES_H10 128
 
 /* For BUCKET_SWEEP == 1, enabling the dictionary lookup makes compression
    a little faster (0.5% - 1%) and it compresses 0.15% better on small text
@@ -546,6 +239,7 @@ static BROTLI_INLINE void FN(StitchToPreviousBlock)(HashToBinaryTree* self,
 #define HASHER() H2
 #define BUCKET_BITS 16
 #define BUCKET_SWEEP 1
+#define HASH_LEN 5
 #define USE_DICTIONARY 1
 #include "./hash_longest_match_quickly_inc.h"  /* NOLINT(build/include) */
 #undef BUCKET_SWEEP
@@ -567,6 +261,7 @@ static BROTLI_INLINE void FN(StitchToPreviousBlock)(HashToBinaryTree* self,
 #define USE_DICTIONARY 1
 #include "./hash_longest_match_quickly_inc.h"  /* NOLINT(build/include) */
 #undef USE_DICTIONARY
+#undef HASH_LEN
 #undef BUCKET_SWEEP
 #undef BUCKET_BITS
 #undef HASHER
@@ -641,12 +336,24 @@ static BROTLI_INLINE void FN(StitchToPreviousBlock)(HashToBinaryTree* self,
 
 #undef BUCKET_BITS
 
+#define HASHER() H54
+#define BUCKET_BITS 20
+#define BUCKET_SWEEP 4
+#define HASH_LEN 7
+#define USE_DICTIONARY 0
+#include "./hash_longest_match_quickly_inc.h"  /* NOLINT(build/include) */
+#undef USE_DICTIONARY
+#undef HASH_LEN
+#undef BUCKET_SWEEP
+#undef BUCKET_BITS
+#undef HASHER
+
 #undef FN
 #undef CAT
 #undef EXPAND_CAT
 
 #define FOR_GENERIC_HASHERS(H) H(2) H(3) H(4) H(5) H(6) H(7) H(8) H(9) \
-                               H(40) H(41) H(42)
+                               H(40) H(41) H(42) H(54)
 #define FOR_ALL_HASHERS(H) FOR_GENERIC_HASHERS(H) H(10)
 
 typedef struct Hashers {
@@ -662,8 +369,8 @@ static BROTLI_INLINE void InitHashers(Hashers* self) {
 }
 
 static BROTLI_INLINE void DestroyHashers(MemoryManager* m, Hashers* self) {
-  if (self->h10) CleanupH10(m, self->h10);
-#define CLEANUP_(N) BROTLI_FREE(m, self->h ## N)
+#define CLEANUP_(N) if (self->h ## N) CleanupH ## N(m, self->h ## N);   \
+                    BROTLI_FREE(m, self->h ## N);
   FOR_ALL_HASHERS(CLEANUP_)
 #undef CLEANUP_
 }
@@ -686,7 +393,12 @@ static BROTLI_INLINE void HashersSetup(
     default: break;
   }
   if (BROTLI_IS_OOM(m)) return;
-  if (type == 10) InitializeH10(self->h10);
+  switch (type) {
+#define INITIALIZE_(N) case N: InitializeH ## N(self->h ## N); break;
+    FOR_ALL_HASHERS(INITIALIZE_);
+#undef INITIALIZE_
+    default: break;
+  }
   HashersReset(self, type);
 }
 
@@ -718,6 +430,24 @@ static BROTLI_INLINE void HashersPrependCustomDictionary(
     default: break;
   }
   if (BROTLI_IS_OOM(m)) return;
+}
+
+static BROTLI_INLINE void InitOrStitchToPreviousBlock(
+    MemoryManager* m, Hashers* self, const uint8_t* data, size_t mask,
+    const BrotliEncoderParams* params, size_t position,
+    size_t bytes, BROTLI_BOOL is_last) {
+  int hasher_type = ChooseHasher(params);
+  switch (hasher_type) {
+#define INIT_(N)                                                              \
+    case N:                                                                   \
+      InitH ## N(m, self->h ## N, data, params, position, bytes, is_last);    \
+      if (BROTLI_IS_OOM(m)) return;                                           \
+      StitchToPreviousBlockH ## N(self->h ## N, bytes, position, data, mask); \
+    break;
+    FOR_ALL_HASHERS(INIT_)
+#undef INIT_
+    default: break;
+  }
 }
