+++ enc/memory.h
@@ -39,7 +39,8 @@ BROTLI_INTERNAL void BrotliInitMemoryManager(
     void* opaque);
 
 BROTLI_INTERNAL void* BrotliAllocate(MemoryManager* m, size_t n);
-#define BROTLI_ALLOC(M, T, N) ((T*)BrotliAllocate((M), (N) * sizeof(T)))
+#define BROTLI_ALLOC(M, T, N)                               \
+  ((N) ? ((T*)BrotliAllocate((M), (N) * sizeof(T))) : NULL)
 
 BROTLI_INTERNAL void BrotliFree(MemoryManager* m, void* p);
 #define BROTLI_FREE(M, P) { \
