+++ include/brotli/port.h
@@ -36,7 +36,7 @@
 
 #if defined(BROTLI_BUILD_MODERN_COMPILER)
 #define BROTLI_MODERN_COMPILER 1
-#elif BROTLI_GCC_VERSION > 300 || BROTLI_ICC_VERSION >= 1600
+#elif BROTLI_GCC_VERSION >= 304 || BROTLI_ICC_VERSION >= 1600
 #define BROTLI_MODERN_COMPILER 1
 #else
 #define BROTLI_MODERN_COMPILER 0
