+++ setup.py
@@ -170,12 +170,14 @@ EXT_MODULES = [
             'dec/huffman.c',
             'dec/state.c',
             'enc/backward_references.c',
+            'enc/backward_references_hq.c',
             'enc/bit_cost.c',
             'enc/block_splitter.c',
             'enc/brotli_bit_stream.c',
             'enc/cluster.c',
             'enc/compress_fragment.c',
             'enc/compress_fragment_two_pass.c',
+            'enc/dictionary_hash.c',
             'enc/encode.c',
             'enc/entropy_encode.c',
             'enc/histogram.c',
@@ -199,6 +201,7 @@ EXT_MODULES = [
             'dec/streams.h',
             'dec/transform.h',
             'enc/backward_references.h',
+            'enc/backward_references_hq.h',
             'enc/backward_references_inc.h',
             'enc/bit_cost.h',
             'enc/bit_cost_inc.h',
@@ -217,6 +220,7 @@ EXT_MODULES = [
             'enc/fast_log.h',
             'enc/find_match_length.h',
             'enc/hash.h',
+            'enc/hash_to_binary_tree_inc.h',
             'enc/hash_longest_match_inc.h',
             'enc/hash_longest_match_quickly_inc.h',
             'enc/histogram.h',
